Move mock data logic to mock folder and minor changes
